requirements.txt
tensorflow-cpu
keras
numpy
pandas
matplotlib
scikit-learn
opencv-python
keras-tuner
scikeras